requirements.txt
coverage==4.4.1
Django==1.11.2
django-filter==1.0.4
django-nose==1.4.4
djangorestframework==3.6.3
factory-boy==2.8.1
Faker==0.7.15
flake8==3.3.0
Markdown==2.6.8
mccabe==0.6.1
nose==1.3.7
pbr==2.0.0
pep257==0.7.0
pycodestyle==2.3.1
pyflakes==1.5.0
python-dateutil==2.6.0
pytz==2017.2
six==1.10.0
stevedore==1.21.0
virtualenv==15.1.0
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.2
